libs: extract shared storage-quota and file-picker helpers
Move the persistent-storage request, the remaining-quota estimate and the disk
file-picker out of the GeoTIFF overlay modules into standalone helpers, so the
upcoming custom-tile-provider import shares one implementation of each instead
of duplicating the navigator.storage and hidden-input plumbing.
